backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.29.7
python-dotenv==1.0.0
pydantic-settings==2.1.0
Pillow==11.0.0
opencv-python==4.10.0.84
numpy==2.1.0